refactor(mcp): divide each framework in different files and separate the logic for STDio and SSE
- Add AgnoMCPServer, GoogleMCPServer, LangchainMCPServer,
  LlamaIndexMCPServer, OpenAIMCPServer, and SmolagentsMCPServer
  classes to manage MCP connections for different frameworks.
- Update .pre-commit-config.yaml to include 'any-agent[dev,all]'.
- Refactor MCPServerBase to handle tool setup based on parameters.
- Enhance test coverage for MCP server functionalities.
