+++ .pre-commit-config.yaml
@@ -30,10 +30,7 @@ repos:
           - types-pyyaml
           - pytest
           - pydantic
-          - openinference-instrumentation-langchain
-          - openinference-instrumentation-llama-index
-          - openinference-instrumentation-smolagents
-          - openinference-instrumentation-openai-agents>=0.1.5
+          - any-agent[dev,all]
 
   - repo: https://github.com/codespell-project/codespell
     rev: v2.3.0
